Support for reading/writing container metadata
Adds the :clear_existing_metadata option which, when set to true,
will add flags to ffmpeg to remove existing metadata on the media
container.

Adds the :metadata option which accepts either a String containing
the key="value" pair you wish to set on the container, or a Hash
of key-value options which will be set on the container.

Tests included.