Allow for custom HEAD request URLs
When using an S3 presigned URL, you cannot obtain a URL for a private object that allows
for both a GET and a HEAD request on it.

Because of this, you need to:

  head_url = object.presigned_url(:head)
  get_url = object.presigned_url(:get)

...independently of each other.

This means that you cannot get media information with this gem.

To _allow_ for this, I've added an option to pass in a custom head_url: argument to allow
for this separation.

In this manner, you can now do this:

  movie = FFMPEG::Movie.new(object.presigned_url(:get), head_url: object.presigned_url(:head))

And obtain your information successfully.